Updated importation of traffic generator variables
The variables that traffic generator imports from "parameters.py"
are now all imported in the main class of the environment and passed
as input parameters to the constructor class of "traffic_generator".
Also added the rates at which each application generates petitions
to their definition (in T = x ms).
Slight update to the histograms as they now overlap datasets but do
not fill the bars.