Core manager and traffic generator debugging
Fixed issue where time slots could remain in queue with start
times under 0.
Fixed issue where the exponential distribution was generating
the arrival time of petitions as the inverse of the mean instead
of the mean. With that, the rate at which each application is
generated has been increased and the cores have had their
resources reduced (to increase the relative load).