Updated route precalculation
The environment now calculates only the necessary routes from
the vehicles defined in the network to other nodes that can
process their applications. As of the current version, the
the only restriction on processing is that vehicles cannot
processes applications from other vehicles.
Still adjusting the size of graphs by trial and error.