Polishing
The comments in all files have been revised and updated
accordingly.
The environment now checks if there is only one cloud
node defined in the network when initializing.
Defined a variable in the environment which defines the
precision of the times used in the simulator.
Updated the graph of histograms with subplots to better
space each subplot.