Initial commit of main script
The script "main.py" has been included for tracking. It is
currently not very useful, and the logging of certain
metrics is incorrect. Future commits are required.
In "offloader.py" the training and testing function now
returns the values of metrics as a dictionary.
Changed the spacing of lines in "graph_creator".