Implemented random variation of processing time
The core manager now functions with estimated times for arrival,
processing and returning petitions (or reservations).
When updating the queue it will generate random numbers with a
normal distribution within defined limits to modify the processing
times of reservations and update the queue accordingly.
The measuring of what reservations where deleted due to randomness
is not implemented!!
The reward calculation needs to be updated to follow a posteriori
methodology!!