Implemented reward a posteriori
The environment now calculates the reward in each time step
as a sum of an instant reward (a priori) and an accumulated
one (a posteriori). The reward for applications queued in
simulated cores won't be received by the agent until the
processing of that application has ended at the correponding
core.
Updated the data collection for metrics to ajust for this
change.
Debbuging of "graph_creator.py"'s subplot histogram continues.
There is an issue with the insertion of the dashed line for
the threshold.