Optimized "core_manager.py" and added shared links BW approximation
The core manager now only updates one vehicle's cores at each time
step. This update includes the time that where accumulated due to
not updating at every time step. The update vehicle is the one that
has to appear in the observation at that time step.
NOTE: This means that rewards for reservations in vehicles can be
delayed.
The bitrates of the shared wireless links (between RSUs and
vehicles) are now the worst case scenario fraction of the total
bitrate at all times.