Bug fixing and apps' definided rates changed
Fixed issue with changing the total number of vehicles in
the environment. The environment now has a defined set function
for this purpose.
Changed the rates at which each app generates petition to be in
line with our definitions (the rates where incresed to compensate
for having only one vehicle).