Reservation without planning and bug fixing
Implemented into "core_manager.py" a function for making
reservations without planning. This means that the reservations
will simply be placed at the end of the shortest queue of the
node (i.e. the queue which has the earliest starting last time
slot).
Created a new function in "core_manager.py" called "reserve()"
used for reserving a petition once the slot to do so has been
calculated. This prevents the functions for reserving with or
without planning repeat their code.
Fixed a bug which made each subsequent reset of the environment
reduce the link rate of the wireless RSU links more that the
last.
Fixed a bug where the delays where being calculated in seconds
but treated as miliseconds (the order of magnitude was wrong).