Simulation functions improved flexibility and renamed main
In "offloader.py" the function used to define the scenario, train
the agents and test then has been separated into three functions
that each individually accomplishes one of those functions.
The script "main.py" has been renamed to "parametric_sim.py" as
it isn't always the main script.
In "parametric_sim.py" defined function to launch a parametric
simulation varying the vehicles in the network.
Added a function to the environment to change the error variation
used to add noise to the processing times at the cores.