Debugging and added new heuristic
The work in progress heuristic for max distance processing is
finished. It asks the environment what the estimated delay will
be for a specific action for the next petition (without accounting
for queueing), and chooses the first acceptable action.
In the environments the selection of the path for an action, and
the calculation of the estimated delays (without queueing) for
that path and action are now independent functions.