Improved the logging and storing of simulation data
The necessary directories where figures are store after
simulations are now created automatically if not existing.
This prevents an error from ocurring if the file path is
incorrect.
Added logging to the training and testing functions so that
a log file stores what they print on console automatically.
The figures are now always created and store too.