Observation expansion, added variation coeficient and bug fixing
The observation is now comprised of a value for each queue that
indicates how full it is, four values which represent next
application parameters relative to the maximum in the set and,
a prediction of how much margin to the maximum latency will remain
if each action is taken (one value per action normalized to the
maximum latency). The latter was added.
The error variation estimation is now inputed as the variation
coeficient of a normal distribution. This allows the standard
deviation of each application's processing time to be dependent
and proportional to its estimated processing time (i.e. the
variance is proportionaly greater the greater the mean).
Fixed a bug which raised an error when a histogram was going to
be created while no samples existed. It now returns an empty
histogram.