Created multiple environmets with different scenarios
There are now four versions of the implemented environment, with each
having a slight difference in its functioning. These environments are:
offload-planning-v0, offload-noplanning-v0, offload-planning-v1 and
offload-noplanning-v1.
The version shows whether the rewards
are calculated with the original design (v0) or the new one (v1). The
original penalizes the failure to process and the failure to process
in time with an added penalty proportional to the time over the latency.
The new desing gives a positive +1 for successful processing and -1 for
unsuccessful processing, with this reward being multiplied by a benefit
value now defined for each application in the datasets.
Planning or noplanning simply refers to whether the queues can use
available time slots in between reservation or not.