Corrected environment for branching networks and bug fixes
The environments now inform the agents of what vehicle node
generated the current petition. It is done by a binary vector
of length the number of vehicle nodes with all 0s except for
one 1, appended to the observation.
For the action distribution, joined the probabilities of all
vehicle nodes.
The render function of the environments now includes information
on the vehicle node generating the petition.
The core manager now treats a detected imprecision as a warning
instead of an error. It will print on console some information
about the afected queue.
Reduced the printing of information about the environment during
training to one every 100 time steps (was once every 10).
Fixed a bug where the vehicle index calculated for the core
manager for the reservation was always from the first vehicle
node.