Added brached network and fixed bug
A network called "Valladolid's network" has been created. It is
identical to the "Branchless network" but has three RSUs and
three vehicle nodes instead of one of each.
Fixed a bug where the testing benefit for a parametric simulation
wasn't being shown in a figure because the testing success rate
was used instead.