Increased simulation stability
The core manager now considers differences close to the precision
limit as equality. This is done with the precision margin.
Multiple checks for updating queues have been rewritten to avoid
imprecisions causing false results on checks and to repair random
imprecision errors.
Added rounding to the precision limit to all operations involving
floating point number in the core manager.
Fixed a bug where the last slot of a queue could be deleted.