Changed per one to % and best agent based on success rate
The success rate is now saved in percentage instead of per one.
The label for the testing success rate has been changed to
include the correct unit (%).
Changed the way the best agent is selected after testing. It is
now chosen based on the the highest success rate.
Removed optimal reward 0. Is now None.